removed FC_FUNC from IPhreeqc_interface_F.h, fixed padfstring to use int sofar
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@9357 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ IPhreeqc_interface_F.cpp
@@ -28,7 +28,7 @@ f2cstring(char* fstring, size_t len)
 void
 padfstring(char *dest, const char *src, int* len)
 {
-    size_t sofar;
+    int sofar;
 
     for (sofar = 0; (sofar < *len) && (*src != '\0'); ++sofar)
         *dest++ = *src++;
